+++ docsig/_main.py
@@ -30,10 +30,10 @@ def main() -> int:
 
     members = _get_members(paths)
     for module in members:
-        _populate(module.name, module, failures, parser.args.disable)
+        _populate(module.name, module, failures, parser.args)
         for klass in module.classes:
             name = f"{module.name}::{klass.name}"
-            _populate(name, klass, failures, parser.args.disable)
+            _populate(name, klass, failures, parser.args)
 
     _print_failures(failures)